refactor(GithubIssues): separate the IssueList component
